+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Section date timezone in chat.
 - Use custom font for chat compose area.
+- Calling conference from history.
 
 ## 5.0.15 - 2023-04-11